+++ packages/bpk-docs/src/pages/FormsPage/InputContainer.js
@@ -45,6 +45,7 @@ class InputContainer extends Component {
       overrideProps = {
         value: this.state.value,
         onChange: e => this.setState({ value: e.target.value }),
+        onClear: () => this.setState({ value: '' }),
       };
     }
     return <FormComponent {...rest} {...overrideProps} />;